feat: external module for create header implemented
